conversation_service_v3/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LangChain core
langchain-core==0.1.23
langchain-openai==0.0.5
langchain==0.1.6

# OpenAI
openai==1.10.0

# HTTP client
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
